+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:demoflu/demoflu.dart';
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 
 void main() {
   DemoFluApp app = DemoFluApp(title: 'Example', rootMenus: _rootMenus);
@@ -9,8 +9,16 @@ void main() {
 List<DemoMenuItem> _rootMenus = [DemoMenuItem('My page', page: () => MyPage())];
 
 class MyPage extends DemoFluPage {
+  final ValueNotifier<int> _count = ValueNotifier(0);
+
   @override
-  void initialize(BuildContext context) {
-    text(text: 'Hello');
+  PageSections buildSections(BuildContext context) {
+    PageSections sections = PageSections();
+    sections.text(text: 'Hello');
+    sections.widget((context) => ElevatedButton(
+        onPressed: () => _count.value++, child: const Text('Increment')));
+    sections.widget((context) => Text('Count: ${_count.value}'),
+        listenable: _count);
+    return sections;
   }
 }