+++ lib/src/widgets/page_widget.dart
@@ -1,7 +1,6 @@
 import 'package:demoflu/src/menu_item.dart';
 import 'package:demoflu/src/page/page.dart';
-import 'package:demoflu/src/model.dart';
-import 'package:demoflu/src/provider.dart';
+import 'package:demoflu/src/page/page_sections.dart';
 import 'package:demoflu/src/widgets/breadcrumb.dart';
 import 'package:demoflu/src/widgets/section_collection_widget.dart';
 import 'package:flutter/material.dart';
@@ -9,21 +8,34 @@ import 'package:meta/meta.dart';
 
 /// Widget for page.
 @internal
-class DemoFluPageWidget extends StatelessWidget {
-  const DemoFluPageWidget({super.key});
+class DemoFluPageWidget extends StatefulWidget {
+  const DemoFluPageWidget({required this.menuItem, super.key});
 
+  final DemoMenuItem menuItem;
+
+  @override
+  State<StatefulWidget> createState() => DemoFluPageState();
+}
+
+class DemoFluPageState extends State<DemoFluPageWidget> {
+  late DemoFluPage? _page;
+
+  @override
+  void initState() {
+    super.initState();
+    if (widget.menuItem.page != null) {
+      _page = widget.menuItem.page!();
+    }
+  }
+
+  @override
   Widget build(BuildContext context) {
-    DemoFluModel model = DemoFluProvider.modelOf(context);
-
     List<Widget> children = [];
-
-    final DemoMenuItem menuItem = model.selectedMenuItem;
-    if (menuItem.page != null) {
-      final DemoFluPage page = menuItem.page!();
-      page.initialize(context);
-      children.add(BreadcrumbWidget(menuItem: menuItem));
+    if (_page != null) {
+      PageSections sections = _page!.buildSections(context);
+      children.add(BreadcrumbWidget(menuItem: widget.menuItem));
       children.add(SizedBox(height: 24));
-      children.add(SectionCollectionWidget(collection: page));
+      children.add(SectionCollectionWidget(collection: sections));
     }
 
     return SingleChildScrollView(